Putting the code from plugin/parse_ls_colors.py in the colorscheme file. Setting the colorscheme back to top level of the git repo.
